+++ src/features/dashboard/TransactionTableSm/table.scss
