+++ src/features/dashboard/dashboard.css
@@ -1,0 +1,16 @@
+.pie_section {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.pie_wrapper {
+  width: 100%;
+  max-width: 1167px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  margin: 4rem 0px;
+}
