+++ src/features/homepage_contents/home_hero.scss
