+++ src/features/pay-calculator/pay-calculator.css
