+++ src/features/pay-calculator/pay-calculator.js
