+++ src/features/pay-calculator/select-input.css
