+++ src/features/pay-calculator/set-pay-components/Results.js
