+++ src/features/pay-calculator/set-pay-components/Settings.js
