+++ src/features/transactions-page/main-table/FullTable.js
