+++ src/features/transactions-page/main-table/mainTrxTable.scss
