+++ src/styles/partials/_transactionPage.scss
