+++ src/styles/partials/_video-modal.scss
